Respect CancellationToken by ASTVisitor
